pyproject.toml
[tool.poetry]
name = "bciip"
version = "0.1.0"
description = "Bangladesh Continuous Internet Intelligence Platform"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "services"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
minio = "^7.2.3"
celery = "^5.3.6"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
feedparser = "^6.0.11"
beautifulsoup4 = "^4.12.3"
pypdf = "^4.0.0"
fasttext-wheel = "^0.9.2"
numpy = "<2.0"
sentence-transformers = "^2.2.2"
pgvector = "^0.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
